is_nary method + test, gen_random args
Added a method to ParentLattice to check whether a lattice definition corresponds to an n_ary. I added a test for this method. This was needed to allow for an integer argument in the creation of random structures, to avoid the weird format {0:[n]} to input the concentration of a simple binary material.