LASSO on residual scheme and slightly modified CV score
Implemented an strategy for cluster selection which I name "lasso on residual". It gives the possibility of defining a fixed set of clusters and do lasso selection on the remaining ones.
I changed the CV score to 10 percent instead of LOO. It's faster and more stable. I also used the multiprocessing option for the call to cross_val_score