Fix computation of correlations for n-aries and small supercells
When a cluster is wrapped into a supercell, if two cluster points ended up in the same crystal site, the cluster was removed if such sites contained different basis functions (labeled by species numbers). This is wrong, since two cluster functions can be evaluated on the same site. There is no problem with that. So, the pieces of code that discarded such clusters are now removed. This bug was affecting the calculation of cluster correlations in small supercells for n-aries with n>2.