refactor: Orders of magnitude acceleration of energy evaluation in MC and WL
The predict_swap function has been refactored such that the interactions associated to every crystal site are put in a dictionary, in an initialization step. Thus, the evaluation of the energy change does not require scanning the whole cluster orbits. With this modification, the energy evaluation in MC and WL computations is essentially independent of system size. Removed old predict_swap routines