Various improvements in WangLandau
refactor: get_cdos has been largely rewritten to avoid conflicts in large supercells (e.g. log of binom is computed with gamma functions instead of first binom and then log), make it more readable and improve the interface and documentation. A new arguemnt "normalization_type" allows better control of and clarifies normalization.

refactor: calculate_thermodynamic_property has been largely rewritten to avoid problems of out of range numbers in large supercells. Make it more readable and improve interface. Computation of logZ with the help of canonical distributioon, instead of directly Z.

fix: visualization of log(g(E)), the logarithm of the configurational density of states, was incorrectly normalized by the mean instead of the module of the mean.